cleaning after refactor
