+++ core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/components/ema-page-dropzone/ema-page-dropzone.component.ts
@@ -12,56 +12,9 @@ import { DotMessagePipe } from '@dotcms/ui';
 
 import { DotErrorPipe } from './pipes/error/dot-error.pipe';
 import { DotPositionPipe } from './pipes/position/dot-position.pipe';
+import { Row, EmaDragItem } from './types';
 
-import { ActionPayload, PositionPayload, ClientData } from '../../../shared/models';
-
-export interface ContentletArea {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-    payload: ActionPayload;
-}
-
-export interface ClientContentletArea {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-    payload: ClientData;
-}
-
-export interface Container {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-    contentlets: ContentletArea[];
-    payload: ClientData | string;
-}
-
-interface Column {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-    containers: Container[];
-}
-
-export interface EmaDragItem {
-    baseType: string;
-    contentType: string;
-}
-
-export interface Row {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-    columns: Column[];
-}
-
-export type EmaPageDropzoneItem = Row | Column | Container | ContentletArea;
+import { PositionPayload, ClientData } from '../../../shared/models';
 
 @Component({
     selector: 'dot-ema-page-dropzone',
